# Zap-Connect/lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * ZapAuto - WhatsApp Materials Consultation System
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type WhatsappStatusStatus =
  (typeof WhatsappStatusStatus)[keyof typeof WhatsappStatusStatus];

export const WhatsappStatusStatus = {
  disconnected: "disconnected",
  connecting: "connecting",
  connected: "connected",
  qr_ready: "qr_ready",
} as const;

export interface WhatsappStatus {
  connected: boolean;
  phone?: string | null;
  name?: string | null;
  status: WhatsappStatusStatus;
}

export interface QrCodeData {
  qrCode?: string | null;
  status: string;
}

export interface OperationResult {
  success: boolean;
  message: string;
}

export interface WhitelistResponse {
  numbers: string[];
  total: number;
}

export interface AddPhoneRequest {
  phone: string;
}

export interface UploadResult {
  success: boolean;
  message: string;
  rowCount?: number;
  fileName?: string;
}

export interface SpreadsheetStatus {
  loaded: boolean;
  fileName?: string | null;
  rowCount: number;
  lastUpdated?: string | null;
}

export interface PreviewRow {
  materialCode: string;
  description: string;
  opCode: string;
  empenhQty: number;
  planDateStr: string;
  status: string;
  currentStock: number;
  purchaseOrderQty: string;
  expectedArrival: string;
}

export interface SpreadsheetStatsResponse {
  totalRows: number;
  totalMaterials: number;
  totalOps: number;
  loaded: boolean;
  fileName?: string | null;
  lastUpdated?: string | null;
  previewRows: PreviewRow[];
}

export interface MissingMaterial {
  materialCode: string;
  description: string;
  qtdFalta: number;
  purchaseOrderQty?: string | null;
  expectedArrival?: string | null;
}

export interface OpQueryResult {
  opCode: string;
  found: boolean;
  hasMissingMaterials: boolean;
  missingMaterials: MissingMaterial[];
  message?: string;
}

export interface OpUsage {
  opCode: string;
  empenhQty: number;
  planDateStr: string;
}

export interface MaterialQueryResult {
  materialCode: string;
  description?: string;
  currentStock?: number | null;
  purchaseOrderQty?: string | null;
  expectedArrival?: string | null;
  opsUsing: OpUsage[];
  found: boolean;
}

export type QueryHistoryItemQueryType =
  (typeof QueryHistoryItemQueryType)[keyof typeof QueryHistoryItemQueryType];

export const QueryHistoryItemQueryType = {
  op: "op",
  material: "material",
} as const;

export interface QueryHistoryItem {
  id: number;
  queryType: QueryHistoryItemQueryType;
  queryValue: string;
  phoneNumber?: string | null;
  result: string;
  createdAt: string;
}

export interface QueryHistoryResponse {
  items: QueryHistoryItem[];
  total: number;
}

export interface QueryStats {
  totalQueries: number;
  queriesLast24h: number;
  opQueries: number;
  materialQueries: number;
  uniqueUsers: number;
  totalMaterials: number;
  totalOps: number;
  sheetLoaded: boolean;
}

export type UploadSpreadsheetBody = {
  file: Blob;
};

export type GetQueryHistoryParams = {
  limit?: number;
};