# Zap-Connect/lib/api-zod/src/generated/types/materialQueryResult.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * ZapAuto - WhatsApp Materials Consultation System
 * OpenAPI spec version: 0.1.0
 */
import type { OpUsage } from "./opUsage";

export interface MaterialQueryResult {
  materialCode: string;
  description?: string;
  currentStock?: number | null;
  purchaseOrderQty?: string | null;
  expectedArrival?: string | null;
  opsUsing: OpUsage[];
  found: boolean;
}