# Zap-Connect/lib/api-zod/src/generated/types/missingMaterial.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * ZapAuto - WhatsApp Materials Consultation System
 * OpenAPI spec version: 0.1.0
 */

export interface MissingMaterial {
  materialCode: string;
  description: string;
  qtdFalta: number;
  purchaseOrderQty?: string | null;
  expectedArrival?: string | null;
}