# Zap-Connect/lib/api-zod/src/generated/types/previewRow.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * ZapAuto - WhatsApp Materials Consultation System
 * OpenAPI spec version: 0.1.0
 */

export interface PreviewRow {
  materialCode: string;
  description: string;
  opCode: string;
  empenhQty: number;
  planDateStr: string;
  status: string;
  currentStock: number;
  purchaseOrderQty: string;
  expectedArrival: string;
}