# Zap-Connect/lib/api-zod/src/generated/types/qrCodeData.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * ZapAuto - WhatsApp Materials Consultation System
 * OpenAPI spec version: 0.1.0
 */

export interface QrCodeData {
  qrCode?: string | null;
  status: string;
}