# Zap-Connect/lib/api-zod/src/generated/types/spreadsheetStatsResponse.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * ZapAuto - WhatsApp Materials Consultation System
 * OpenAPI spec version: 0.1.0
 */
import type { PreviewRow } from "./previewRow";

export interface SpreadsheetStatsResponse {
  totalRows: number;
  totalMaterials: number;
  totalOps: number;
  loaded: boolean;
  fileName?: string | null;
  lastUpdated?: string | null;
  previewRows: PreviewRow[];
}