# Zap-Connect/lib/api-zod/src/generated/types/whatsappStatus.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * ZapAuto - WhatsApp Materials Consultation System
 * OpenAPI spec version: 0.1.0
 */
import type { WhatsappStatusStatus } from "./whatsappStatusStatus";

export interface WhatsappStatus {
  connected: boolean;
  phone?: string | null;
  name?: string | null;
  status: WhatsappStatusStatus;
}